Rename spacetime-parser to spacetime-lang
