+++ Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "spacetime-parser"
+name = "spacetime-lang"
 version = "0.1.0"
 edition = "2021"
