+++ go/parser/.gitignore
@@ -1,1 +1,1 @@
-libspacetime_parser.a
+libspacetime_lang.a
