+++ js/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@spacetimexyz/parser",
+  "name": "@spacetimexyz/lang",
   "version": "0.1.3",
   "description": "",
   "types": "./node/index.d.ts",
@@ -35,6 +35,6 @@
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/spacetimehq/spacetime-parser.git"
+    "url": "git+https://github.com/spacetimehq/spacetime-lang.git"
   }
 }
